Added: enemyService for Stalking the opponent
