+++ Cluster/Cluster.py
@@ -8,6 +8,7 @@ from lux.game_constants import GAME_CONSTANTS
 from helperFunctions.helper_functions import *
 from Map.mapService import get_cell_neighbours_four, get_nearest_position
 from Units.unitsService import get_unit_by_id
+from Enemy.enemyService import get_enemy_tiles
 
 from Resources.resourceService import get_resources_from_cells
